% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RabixTool-class.R
\docType{class}
\name{RabixTool-class}
\alias{RabixTool}
\alias{RabixTool-class}
\title{Rabix CommandLineTool Class}
\description{
Rabix subclass for CommandLineTool used by rabix.org or sbg
platform. \code{RabixTool} class extends \code{CommandLineTool}
with more fields.
}
\section{Fields}{

\describe{
\item{\code{context}}{[character] by default:
"https://github.com/common-workflow-language/common-workflow-language/blob/draft-1/specification/tool-description.md"}

\item{\code{owner}}{[list] a list of owner names.}

\item{\code{contributor}}{[list] a list of contributor names.}
}}
\section{Methods}{

\describe{
\item{\code{toList(...)}}{Convert object to a list of simple data types}
}}

