% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/99-RabixTool-class.R
\docType{class}
\name{RabixTool-class}
\alias{RabixTool}
\alias{RabixTool-class}
\title{Rabix CommandLineTool Class}
\description{
Rabix subclass for CommandLineTool used by rabix.org or sbg
platform. \code{RabixTool} class extends \code{CommandLineTool}
with more fields.
}
\section{Fields}{

\describe{
\item{\code{context}}{[character] by default:
"https://github.com/common-workflow-language/common-workflow-language/blob/draft-1/specification/tool-description.md"}

\item{\code{owner}}{[list] a list of owner names.}

\item{\code{contributor}}{[list] a list of contributor names.}
}}
\section{Methods}{

\describe{
\item{\code{toList(...)}}{Convert object to a list of simple data types}
}}
\section{other fields}{

\describe{
\item{\code{cpu}}{cpu 0 or 1, for any value >1 will be converted to 1L, passed to CpuRequirement}
\item{\code{mem}}{Positive integer. Passed to MemRequirement.}
\item{\code{dockerPull}}{[character] Get a Docker image using
docker pull}

\item{\code{dockerLoad}}{[character] Specify a HTTP URL from which
to download a Docker image using docker load.}

\item{\code{dockerFile}}{[character] Supply the contents of a
Dockerfile which will be build using docker build.}

\item{\code{dockerImageId}}{[character] The image id that will be
used for docker run. May be a human-readable image name or the
image identifier hash. May be skipped if dockerPull is specified,
in which case the dockerPull image id will be used.}

\item{\code{dockerOutputDirectory}}{ [character] Set the designated
output directory to a specific location inside the Docker
container.}}
}
\examples{
ipl <- IPList(
    InPar(id = "bam",
          type = "File",
          label = "Bam file",
          description = "Input bam file",
          position = 1L,
          separate = TRUE),
    InPar(id = "level",
          type = "Integer",
          label = "Compression Level",
          description = "Set compression level, from 0 (uncompressed) to 9 (best)",
          position = 2L),
    InPar(id = "prefix",
          type = "String",
          label = "Prefix",
          description = "Write temporary files to PREFIX.nnnn.bam",
          position = 3L)
)
opl <- OPList(OutPar(
        id = "sorted",
        type = "File",
        glob = "*.bam"
))
rbx <- RabixTool(id = "samtools-sort",
                label = "Samtools sort subcommand",
                description = "Samtools sort: sort bam into sorted bam : )",
                dockerPull = "tengfei/samtools:v1.2",
                cpu = 2, mem = 202,
                baseCommand = "samtools sort",
                arguments = "out.bam",
                inputs = ipl,
                 outputs = opl)
rbx$toJSON()
rbx$toJSON(pretty = TRUE)
}

